This change fixes build breakage on macOS due to refreshes of
swift-llvm, swift-clang and swift in the stable-next and master-next
branches.

The test run still needs to be verified, as does the Linux build
and test run.

apple-llvm-split-commit: 8da04a2feb924b80b1402df264b223c807e4fb84
apple-llvm-split-dir: lldb/

+++ llvm/include/llvm/ADT/Hashing.h
@@ -52,7 +52,6 @@
 #include <algorithm>
 #include <cassert>
 #include <cstring>
-#include <iterator>
 #include <string>
 #include <utility>
 
@@ -632,7 +631,8 @@ inline hash_code hash_integer_value(uint64_t value) {
 template <typename T>
 typename std::enable_if<is_integral_or_enum<T>::value, hash_code>::type
 hash_value(T value) {
-  return ::llvm::hashing::detail::hash_integer_value(value);
+  return ::llvm::hashing::detail::hash_integer_value(
+      static_cast<uint64_t>(value));
 }
 
 // Declared and documented above, but defined here so that any of the hashing
